backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy[asyncio]==2.0.23
aiopg==1.3.4
asyncpg==0.29.0
alembic==1.12.1
pydantic[email]==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
slowapi==0.1.9
python-dotenv==1.0.0
cors-middleware==0.1.0
